ai-service-agents/requirements.txt
fastapi
uvicorn[standard]
phidata
google-generativeai
google-genai
python-multipart
duckduckgo-search
python-dotenv
pydantic
pillow
phi
httpx